requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
pyjwt==2.8.0
bcrypt==4.0.1
Werkzeug==2.2.0